+++ build.ps1
@@ -44,4 +44,4 @@ Write-Host "NuGet Packages creation"
 #dotnet pack ./src/NEventStore.Persistence.MongoDB/NEventStore.Persistence.MongoDB.Core.csproj --no-build -c $configuration -o $artifacts -p:NuspecFile="" -p:NuspecProperties="pippo=$configuration;version=$nugetversion"
 
 #Write-Host nuget pack ./src/.nuget/NEventStore.Persistence.MongoDB.nuspec -properties "version=$nugetversion;configuration=$configuration"
-nuget pack ./src/.nuget/NEventStore.Persistence.MongoDB.nuspec -properties "version=$nugetversion;configuration=$configuration" -OutputDirectory $artifacts -Symbols
+nuget pack ./src/.nuget/NEventStore.Persistence.MongoDB.nuspec -properties "version=$nugetversion;configuration=$configuration" -OutputDirectory $artifacts -Symbols -SymbolPackageFormat snupkg