+++ src/NEventStore.Persistence.MongoDB.Tests/AcceptanceTestMongoPersistenceFactory.cs
@@ -3,6 +3,7 @@
     using global::MongoDB.Bson.Serialization.Serializers;
     using global::MongoDB.Bson.Serialization;
     using NEventStore.Serialization;
+    using global::MongoDB.Bson;
 #if MSTEST
     using global::MongoDB.Driver;
 #endif
@@ -12,12 +13,17 @@
 
         static AcceptanceTestMongoPersistenceFactory() {
             // MongoDb serialization changed
+            // MongoDb 3.0.0 GUID serialization changed
+            BsonSerializer.RegisterSerializer(new GuidSerializer(GuidRepresentation.CSharpLegacy));
+            // when serializing guid in a Dictionary<object, object> take a look at the comment here:
+            // https://jira.mongodb.org/browse/CSHARP-4987?jql=text%20~%20%22GuidRepresentation%20dictionary%22
+            // it seems you also need to configure the ObjectSerializer
             // What Object Types Can Be Serialized?
-            // https://www.mongodb.com/docs/drivers/csharp/current/faq/#what-object-types-can-be-serialized-
-            BsonSerializer.RegisterSerializer(new ObjectSerializer(ObjectSerializer.AllAllowedTypes));
+            BsonSerializer.RegisterSerializer(new ObjectSerializer(
+                BsonSerializer.LookupDiscriminatorConvention(typeof(object)), GuidRepresentation.CSharpLegacy, ObjectSerializer.AllAllowedTypes));
         }
 
-        public AcceptanceTestMongoPersistenceFactory(MongoPersistenceOptions options = null)
+        public AcceptanceTestMongoPersistenceFactory(MongoPersistenceOptions? options = null)
             : base(
                 GetConnectionString,
                 new DocumentObjectSerializer(),
@@ -30,7 +36,7 @@
         }
 
         internal static string GetConnectionString() {
-            string connectionString = Environment.GetEnvironmentVariable(EnvVarConnectionStringKey, EnvironmentVariableTarget.Process);
+            var connectionString = Environment.GetEnvironmentVariable(EnvVarConnectionStringKey, EnvironmentVariableTarget.Process);
 
             if (connectionString == null) {
                 string message = string.Format(