+++ src/NEventStore.Persistence.MongoDB.Tests/AcceptanceTests/MongoCommitCustomSerializationTests.cs
@@ -1,12 +1,9 @@
-﻿using MongoDB.Bson;
-using MongoDB.Bson.Serialization;
+﻿using MongoDB.Bson.Serialization;
 using MongoDB.Bson.Serialization.Serializers;
 using NEventStore.Persistence.AcceptanceTests;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using FluentAssertions;
 using NEventStore.Persistence.AcceptanceTests.BDD;
 #if MSTEST
@@ -22,17 +19,17 @@ using NUnit.Framework;
 
 namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
 {
-#warning due to problems with the xUnit test runner (it does not skip the tests marked with Fact(Skip="...") in the build server) we have actually disabled these tests commenting out the [Fact] attrubute, to run them you need to add it back manually
-
     /// <summary>
     /// the problem here is that this is 'static', no way to change it once it defined, so the tests need to be run
     /// manually :(
     /// or I need to implement a fully featured serializer specifically designed for testing
+    /// or I reset the ClassMap (maybe using reflection and re-init it again)
     /// </summary>
     internal static class MapMongoCommit
     {
         public static void MapMongoCommit_Header_as_Document()
         {
+            BsonClassMapExtensions.UnregisterClassMap<MongoCommit>();
             if (!BsonClassMap.IsClassMapRegistered(typeof(MongoCommit)))
             {
                 BsonClassMap.RegisterClassMap<MongoCommit>(cm =>
@@ -53,21 +50,19 @@ namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
         // this is the default behavior
         public static void MapMongoCommit_Header_as_ArrayOfArray()
         {
+            BsonClassMapExtensions.UnregisterClassMap<MongoCommit>();
         }
     }
 
-    /// <summary>
-    /// Be carefull! this test will fail with 'Catastrophic Failure' and Visual Studio IDE will report this as not run.
-    /// </summary>
 #if MSTEST
     [TestClass]
 #endif
     public class When_serializing_headers_as_Document_and_a_commit_header_has_a_name_that_contains_a_period : PersistenceEngineConcern
     {
-        // private ICommit _persisted;
-        private string _streamId;
+        private ICommit? _persisted;
+        private string? _streamId;
 
-        private Exception _thrown;
+        private Exception? _thrown;
 
         public When_serializing_headers_as_Document_and_a_commit_header_has_a_name_that_contains_a_period()
         {
@@ -88,25 +83,22 @@ namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
                     1,
                     DateTime.Now,
                     new Dictionary<string, object> { { "key.1", "value" } },
-                    new List<EventMessage> { new EventMessage { Body = new NEventStore.Persistence.AcceptanceTests.ExtensionMethods.SomeDomainEvent { SomeProperty = "Test" } } });
+                    [new() { Body = new NEventStore.Persistence.AcceptanceTests.ExtensionMethods.SomeDomainEvent { SomeProperty = "Test" } }]);
                 Persistence.Commit(attempt);
             });
 
-            // _persisted = Persistence.GetFrom(_streamId, 0, int.MaxValue).First();
+            _persisted = Persistence.GetFrom(_streamId, 0, int.MaxValue).First();
         }
 
-        // Enable this test manually, it does not get skipped in the build server causing the build to fail
-        // [Fact(Skip = "Run it Manually")]
-#if NUNIT
         [Fact]
-        [Explicit("Run as Standalone due to MongoDb mapping configuration being static")]
-#endif
-        public void Should_throw_serialization_exception_due_to_invalid_key()
+        public void Should_correctly_deserialize_headers()
         {
-            // _persisted.Headers.Keys.ShouldContain("key.1");
+            // with previous drivers this resulted in an error
+            // _thrown.Should().BeOfType<BsonSerializationException>();
+            // _thrown.Message.Should().Contain("key.1");
 
-            _thrown.Should().BeOfType<BsonSerializationException>();
-            _thrown.Message.Should().Contain("key.1");
+            _thrown.Should().BeNull();
+            _persisted!.Headers.Keys.Should().Contain("key.1");
         }
     }
 
@@ -115,8 +107,8 @@ namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
 #endif
     public class When_serializing_headers_as_Document_and_a_commit_header_has_a_valid_name : PersistenceEngineConcern
     {
-        private ICommit _persisted;
-        private string _streamId;
+        private ICommit? _persisted;
+        private string? _streamId;
 
         public When_serializing_headers_as_Document_and_a_commit_header_has_a_valid_name()
         {
@@ -132,7 +124,7 @@ namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
                 1,
                 DateTime.Now,
                 new Dictionary<string, object> { { "key", "value" } },
-                new List<EventMessage> { new EventMessage { Body = new NEventStore.Persistence.AcceptanceTests.ExtensionMethods.SomeDomainEvent { SomeProperty = "Test" } } });
+                [new() { Body = new NEventStore.Persistence.AcceptanceTests.ExtensionMethods.SomeDomainEvent { SomeProperty = "Test" } }]);
             Persistence.Commit(attempt);
         }
 
@@ -141,15 +133,11 @@ namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
             _persisted = Persistence.GetFrom(_streamId, 0, int.MaxValue).First();
         }
 
-        // Enable this test manually, it does not get skipped in the build server causing the build to fail
-        // [Fact(Skip = "Run it Manually")]
-#if NUNIT
         [Fact]
-        [Explicit("Run as Standalone due to MongoDb mapping configuration being static")]
-#endif
         public void Should_correctly_deserialize_headers()
         {
-            _persisted.Headers.Keys.Should().Contain("key");
+            _persisted.Should().NotBeNull();
+            _persisted!.Headers.Keys.Should().Contain("key");
         }
     }
 
@@ -158,12 +146,13 @@ namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
 #endif
     public class When_serializing_headers_as_ArrayOfArrays_and_a_commit_header_has_a_name_that_contains_a_period : PersistenceEngineConcern
     {
-        private ICommit _persisted;
-        private string _streamId;
+        private ICommit? _persisted;
+        private string? _streamId;
 
         public When_serializing_headers_as_ArrayOfArrays_and_a_commit_header_has_a_name_that_contains_a_period()
         {
             // the default is ArrayOfArray defined using an attribute.
+            MapMongoCommit.MapMongoCommit_Header_as_ArrayOfArray();
         }
 
         protected override void Context()
@@ -175,7 +164,7 @@ namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
                 1,
                 DateTime.Now,
                 new Dictionary<string, object> { { "key.1", "value" } },
-                new List<EventMessage> { new EventMessage { Body = new NEventStore.Persistence.AcceptanceTests.ExtensionMethods.SomeDomainEvent { SomeProperty = "Test" } } });
+                [new() { Body = new NEventStore.Persistence.AcceptanceTests.ExtensionMethods.SomeDomainEvent { SomeProperty = "Test" } }]);
             Persistence.Commit(attempt);
         }
 
@@ -187,7 +176,99 @@ namespace NEventStore.Persistence.MongoDB.Tests.AcceptanceTests
         [Fact]
         public void Should_correctly_deserialize_headers()
         {
-            _persisted.Headers.Keys.Should().Contain("key.1");
+            _persisted.Should().NotBeNull();
+            _persisted!.Headers.Keys.Should().Contain("key.1");
+        }
+    }
+
+    // guid serialization tests (inside headers)
+
+#if MSTEST
+    [TestClass]
+#endif
+    public class When_serializing_headers_as_Document_and_a_commit_header_has_a_Guid : PersistenceEngineConcern
+    {
+        private ICommit? _persisted;
+        private string? _streamId;
+        private readonly Guid _guid = Guid.NewGuid();
+
+        private Exception? _thrown;
+
+        public When_serializing_headers_as_Document_and_a_commit_header_has_a_Guid()
+        {
+            MapMongoCommit.MapMongoCommit_Header_as_Document();
+        }
+
+        protected override void Context()
+        { }
+
+        protected override void Because()
+        {
+            _thrown = Catch.Exception(() =>
+            {
+                _streamId = Guid.NewGuid().ToString();
+                var attempt = new CommitAttempt(_streamId,
+                    2,
+                    Guid.NewGuid(),
+                    1,
+                    DateTime.Now,
+                    new Dictionary<string, object> { { "guid", _guid } },
+                    [new() { Body = new NEventStore.Persistence.AcceptanceTests.ExtensionMethods.SomeDomainEvent { SomeProperty = "Test" } }]);
+                Persistence.Commit(attempt);
+            });
+
+            Assert.That(_thrown, Is.Null);
+
+            _persisted = Persistence.GetFrom(_streamId, 0, int.MaxValue).First();
+        }
+
+        [Fact]
+        public void Should_correctly_deserialize_headers()
+        {
+            _thrown.Should().BeNull();
+            _persisted!.Headers.Keys.Should().Contain("guid");
+            _persisted.Headers["guid"].Should().Be(_guid);
+        }
+    }
+
+#if MSTEST
+    [TestClass]
+#endif
+    public class When_serializing_headers_as_ArrayOfArrays_and_a_commit_header_has_a_Guid : PersistenceEngineConcern
+    {
+        private ICommit? _persisted;
+        private string? _streamId;
+        private readonly Guid _guid = Guid.NewGuid();
+
+        public When_serializing_headers_as_ArrayOfArrays_and_a_commit_header_has_a_Guid()
+        {
+            // the default is ArrayOfArray defined using an attribute.
+            MapMongoCommit.MapMongoCommit_Header_as_ArrayOfArray();
+        }
+
+        protected override void Context()
+        {
+            _streamId = Guid.NewGuid().ToString();
+            var attempt = new CommitAttempt(_streamId,
+                2,
+                Guid.NewGuid(),
+                1,
+                DateTime.Now,
+                new Dictionary<string, object> { { "guid", _guid } },
+                [new() { Body = new NEventStore.Persistence.AcceptanceTests.ExtensionMethods.SomeDomainEvent { SomeProperty = "Test" } }]);
+            Persistence.Commit(attempt);
+        }
+
+        protected override void Because()
+        {
+            _persisted = Persistence.GetFrom(_streamId, 0, int.MaxValue).First();
+        }
+
+        [Fact]
+        public void Should_correctly_deserialize_headers()
+        {
+            _persisted!.Headers.Keys.Should().Contain("guid");
+            _persisted.Headers["guid"].Should().Be(_guid);
         }
     }
 }